+++ ddtrace/appsec/_iast/_taint_tracking/_context.py
@@ -6,7 +6,7 @@ from ddtrace.appsec._iast._taint_tracking._native.context import debug_taint_map
 from ddtrace.appsec._iast._taint_tracking._native.context import finish_request_context
 from ddtrace.appsec._iast._taint_tracking._native.context import is_in_taint_map
 from ddtrace.appsec._iast._taint_tracking._native.context import start_request_context
-from ddtrace.appsec._iast_context import _get_iast_context_id
+from ddtrace.appsec._shared._iast_context import _get_iast_context_id
 
 
 def _num_objects_tainted_in_request() -> int:
