+++ ddtrace/appsec/_iast/_taint_tracking/_taint_objects.py
@@ -11,8 +11,8 @@ from ddtrace.appsec._iast._taint_tracking import OriginType
 from ddtrace.appsec._iast._taint_tracking import TaintRange
 from ddtrace.appsec._iast._taint_tracking import set_ranges
 from ddtrace.appsec._iast._taint_tracking._taint_objects_base import _taint_pyobject_base
-from ddtrace.appsec._iast_context import _get_iast_context_id
-from ddtrace.appsec._iast_context import _is_iast_taint_source_enabled
+from ddtrace.appsec._shared._iast_context import _get_iast_context_id
+from ddtrace.appsec._shared._iast_context import _is_iast_taint_source_enabled
 from ddtrace.internal.logger import get_logger
