+++ ddtrace/appsec/_iast/reporter.py
@@ -13,7 +13,7 @@ from ddtrace.appsec._iast._utils import _get_source_index
 from ddtrace.appsec._iast.constants import VULN_INSECURE_HASHING_TYPE
 from ddtrace.appsec._iast.constants import VULN_WEAK_CIPHER_TYPE
 from ddtrace.appsec._iast.constants import VULN_WEAK_RANDOMNESS
-from ddtrace.appsec._iast_context import is_iast_request_enabled
+from ddtrace.appsec._shared._iast_context import is_iast_request_enabled
 from ddtrace.internal.logger import get_logger
 from ddtrace.internal.settings import env
 from ddtrace.internal.settings.asm import config as asm_config
