+++ ddtrace/appsec/_shared/_iast_context.py
@@ -6,8 +6,13 @@ from typing import Iterator
 from typing import Optional
 
 
-# AIDEV-NOTE: Keep this outside the _iast package so AppSec can suppress taint
-# sources without initializing IAST or loading its native extensions.
+# AIDEV-NOTE: This module must stay outside the _iast package and must never
+# import from ddtrace (stdlib only). Importing any _iast submodule executes
+# ddtrace/appsec/_iast/__init__.py, which reaches _asm_request_context through
+# _listener -> _iast_request_context -> reporter -> _exploit_prevention.stack_traces
+# and also loads the native taint-tracking extension. Moving these primitives
+# under _iast would therefore re-create the import cycle this module broke, and
+# would make ASM initialize IAST just to suppress taint sources.
 IAST_CONTEXT: contextvars.ContextVar[Optional[int]] = contextvars.ContextVar("iast_var", default=None)
 
 # Keep source suppression separate from IAST_CONTEXT. Clearing the
