+++ tests/appsec/appsec/test_asm_request_context.py
@@ -3,7 +3,7 @@ import logging
 import pytest
 
 from ddtrace.appsec import _asm_request_context
-from ddtrace.appsec._iast_context import _is_iast_taint_source_enabled
+from ddtrace.appsec._shared._iast_context import _is_iast_taint_source_enabled
 from ddtrace.internal._exceptions import BlockingException
 from tests.appsec.utils import asm_context
 from tests.utils import override_global_config
